Major refactoring to simplify and improve universe
